requirements.txt
flask
pandas
transformers
torch
sqlalchemy
pyodbc
openpyxl
azure-storage-blob